requirements.txt
streamlit
streamlit-cookies-controller
cryptography
pandas
requests
python-dotenv